requirements.txt
attrs==21.2.0
base58==2.1.0
bitarray==1.2.2
certifi==2021.10.8
charset-normalizer==2.0.7
cytoolz==0.11.0
eth-abi==2.1.1
eth-account==0.5.6
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
hexbytes==0.2.2
idna==3.3
importlib-metadata==4.8.1
ipfshttpclient==0.6.1
jsonschema==3.2.0
lru-dict==1.1.7
multiaddr==0.0.9
netaddr==0.8.0
parsimonious==0.8.1
protobuf==3.19.0
pycryptodome==3.11.0
pyrsistent==0.18.0
python-dotenv==0.19.1
requests==2.26.0
rlp==2.0.1
six==1.16.0
toolz==0.11.1
typing-extensions==3.10.0.2
urllib3==1.26.7
varint==1.0.2
web3==5.12.2
websockets==8.1
zipp==3.6.0
